+++ src/Vertex.cpp
@@ -12,20 +12,20 @@ namespace GLBase
     vec3 vec3::operator-() const {
         return {-x, -y, -z};
     }
-    vec3 vec3::operator*(double v) const { return {
+    vec3 vec3::operator*(float v) const { return {
         v*x, v*y, v*z
     };}
-    vec3 vec3::operator/(double v) const { return {
+    vec3 vec3::operator/(float v) const { return {
         x/v, y/v, z/v
     };}
-    vec3& vec3::operator/=(double v) {
+    vec3& vec3::operator/=(float v) {
         return *this = *this / v;
     }
-    double vec3::dot(const vec3& other) const {
+    float vec3::dot(const vec3& other) const {
         return x*other.x + y*other.y + z*other.z;
     }
-    double vec3::squaredLength() const { return dot(*this); }
-    double vec3::length() const { return sqrt(dot(*this)); }
+    float vec3::squaredLength() const { return dot(*this); }
+    float vec3::length() const { return sqrt(dot(*this)); }
     vec3 vec3::cross(const vec3& other) const {
         return {
             y * other.z - z * other.y,
